+++ export/browser.js
@@ -41,6 +41,7 @@ import LinkPart from '../src/content-element/part/link-part';
 import SocialFollowPart from '../src/content-element/part/social-follow-part';
 import SocialSharePart from '../src/content-element/part/social-share-part';
 import UrlProviderPart from '../src/content-element/part/url-provider-part';
+import RawPart from '../src/content-element/part/raw';
 import Website from '../src/website/website';
 import PageInclude from '../src/website/page-include';
 import Include from '../src/website/include';
@@ -100,6 +101,7 @@ export {
   SocialFollowPart,
   SocialSharePart,
   UrlProviderPart,
+  RawPart,
   Website,
   PageInclude,
   Include,